refactor: modularize agent implementation
